refact: posts score vars and sorts
