+++ app/leaderboard/leaderboardClient.tsx
@@ -28,31 +28,7 @@ import AirdropModal from "@/components/airdrop/AirdropModal";
 import React from "react";
 import useIsMobile from "@/hooks/useIsMobile";
 import { Name } from "@coinbase/onchainkit/identity";
-import { SkaterData } from "./page";
-
-// interface SkaterData {
-//   id: number;
-//   hive_author: string;
-//   hive_balance: number;
-//   hp_balance: number;
-//   hbd_balance: number;
-//   hbd_savings_balance: number;
-//   has_voted_in_witness: boolean;
-//   eth_address: string;
-//   gnars_balance: number;
-//   gnars_votes: number;
-//   skatehive_nft_balance: number;
-//   max_voting_power_usd: number;
-//   last_updated: string;
-//   last_post: string;
-//   post_count: number;
-//   posts_score: number;
-//   snaps_count: number;
-//   delegated_curator: number;
-//   points: number;
-//   giveth_donations_usd: number;
-//   giveth_donations_amount: number;
-// }
+import { SkaterData } from "@/types/leaderboard";
 
 interface Props {
   skatersData: SkaterData[];
@@ -241,7 +217,7 @@ export default function LeaderboardClient({ skatersData }: Props) {
             b.eth_address !== "0x0000000000000000000000000000000000000000";
 
           if (aHasEth === bHasEth) {
-            return b.post_count - a.post_count; // Same ETH status, sort by activity
+            return b.posts_score - a.posts_score; // Same ETH status, sort by activity
           }
           return aHasEth ? 1 : -1; // Users without ETH rank higher
         case "gnars_balance":
@@ -252,7 +228,7 @@ export default function LeaderboardClient({ skatersData }: Props) {
           // Sort by post count first, then by witness vote presence
           // This shows active users who haven't voted for witness
           if (a.has_voted_in_witness === b.has_voted_in_witness) {
-            return b.post_count - a.post_count; // Same witness status, sort by activity
+            return b.posts_score - a.posts_score; // Same witness status, sort by activity
           }
           return a.has_voted_in_witness ? 1 : -1; // Users without witness vote rank higher
         case "last_updated":
@@ -347,7 +323,7 @@ export default function LeaderboardClient({ skatersData }: Props) {
     {
       key: "posts",
       label: "Posts Score",
-      value: (skater: SkaterData) => skater.post_count,
+      value: (skater: SkaterData) => skater.posts_score,
     },
   ];
 
@@ -377,7 +353,7 @@ export default function LeaderboardClient({ skatersData }: Props) {
     {
       key: "posts",
       label: "Posts Score",
-      value: (skater: SkaterData) => skater.post_count,
+      value: (skater: SkaterData) => skater.posts_score,
     },
     {
       key: "nfts",
