+++ app/leaderboard/page.tsx
@@ -1,29 +1,6 @@
 import { Metadata } from "next";
 import LeaderboardClient from "./leaderboardClient";
-
-export interface SkaterData {
-  id: number;
-  hive_author: string;
-  hive_balance: number;
-  hp_balance: number;
-  hbd_balance: number;
-  hbd_savings_balance: number;
-  has_voted_in_witness: boolean;
-  eth_address: string;
-  gnars_balance: number;
-  gnars_votes: number;
-  skatehive_nft_balance: number;
-  max_voting_power_usd: number;
-  last_updated: string;
-  last_post: string;
-  post_count: number;
-  posts_score: number;
-  snaps_count: number;
-  delegated_curator: number;
-  points: number;
-  giveth_donations_usd: number;
-  giveth_donations_amount: number;
-}
+import { SkaterData } from "@/types/leaderboard";
 
 interface LeaderboardDataResult {
   data: SkaterData[];
