+++ types/airdrop.ts
@@ -1,30 +1,9 @@
-export interface SkaterData {
-  id: number;
-  hive_author: string;
-  hive_balance: number;
-  hp_balance: number;
-  hbd_balance: number;
-  hbd_savings_balance: number;
-  has_voted_in_witness: boolean;
-  eth_address: string;
-  gnars_balance: number;
-  gnars_votes: number;
-  skatehive_nft_balance: number;
-  max_voting_power_usd: number;
-  last_updated: string;
-  last_post: string;
-  post_count: number;
-  points: number;
-  giveth_donations_usd: number;
-  giveth_donations_amount: number;
-}
-
 export type SortOption =
   | "points"
   | "hp_balance" 
   | "hive_balance"
   | "hbd_savings_balance"
-  | "post_count"
+  | "posts_score"
   | "has_voted_in_witness"
   | "gnars_balance"
   | "skatehive_nft_balance"
